backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.6b0
click==8.0.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
